+++ sycl/test-e2e/Graph/Threading/explicit_add_nodes.cpp
@@ -39,7 +39,7 @@ int main() {
 
   std::vector<std::thread> Threads;
   Threads.reserve(NumThreads);
-  for (size_t i = 0; i < NumThreads; ++i) {
+  for (unsigned i = 0; i < NumThreads; ++i) {
     Threads.emplace_back(AddNodesToGraph);
   }
